Make the tool definition clearer
