+++ aworld/agents/gaia/agent.py
@@ -6,10 +6,11 @@ import time
 import traceback
 from typing import Dict, Any, List, Union
 
+from aworld.config.common import Agents
 from aworld.core.agent.base import BaseAgent, AgentFactory
 from aworld.models.utils import tool_desc_transform
 from aworld.config.conf import AgentConfig, ConfigDict
-from aworld.core.common import Observation, ActionModel, Agents
+from aworld.core.common import Observation, ActionModel
 from aworld.logs.util import logger
 from aworld.core.envs.tool_desc import get_tool_desc
 from aworld.agents.gaia.prompts import *
