+++ aworld/agents/travel/searchAgent.py
@@ -3,9 +3,10 @@ import os
 import json
 from typing import Dict, Any, List, Union
 
+from aworld.config.common import Agents, Tools
 from aworld.config.conf import AgentConfig
 from aworld.core.agent.base import AgentFactory, BaseAgent
-from aworld.core.common import Agents, Observation, ActionModel, Tools
+from aworld.core.common import Observation, ActionModel
 from aworld.core.envs.tool_desc import get_tool_desc_by_name
 from aworld.models.utils import tool_desc_transform
 from aworld.core.envs.tool import ToolFactory
@@ -40,19 +41,17 @@ class SearchAgent(BaseAgent):
         # we only use search api tool for example.
         self.tool_desc = tool_desc_transform({Tools.SEARCH_API.value: get_tool_desc_by_name(Tools.SEARCH_API.value)})
 
-    # Step3
-    def name(self) -> str:
-        return Agents.SEARCH.value
-
     def policy(self, observation: Observation, info: Dict[str, Any] = {}, **kwargs) -> Union[
         List[ActionModel], None]:
-        if observation.action_result is not None and len(observation.action_result)!=0 and observation.action_result[0].is_done:
+        if observation.action_result is not None and len(observation.action_result) != 0 and observation.action_result[
+            0].is_done:
             self._finished = True
 
             return [ActionModel(tool_name="[done]", policy_info=observation.content)]
 
         messages = [{'role': 'system', 'content': sys_prompt},
-                    {'role': 'user', 'content': prompt.format(task=observation.content, tool_desc=self.tool_desc) + response_format}]
+                    {'role': 'user',
+                     'content': prompt.format(task=observation.content, tool_desc=self.tool_desc) + response_format}]
 
         llm_result = self.llm.invoke(
             input=messages,
@@ -63,7 +62,7 @@ class SearchAgent(BaseAgent):
 
     def _result(self, data):
 
-        data = json.loads(data.replace("```json","").replace("```",""))
+        data = json.loads(data.replace("```json", "").replace("```", ""))
         actions = data.get("action", [])
         parsed_results = []
 
@@ -79,9 +78,10 @@ class SearchAgent(BaseAgent):
 
                 # 将解析结果存入列表
                 parsed_results.append(ActionModel(tool_name=tool_name,
-                                       action_name=action_name,
-                                       params=params))
+                                                  action_name=action_name,
+                                                  params=params))
         return parsed_results
+
 
 if __name__ == '__main__':
     agentConfig = AgentConfig(
@@ -103,12 +103,10 @@ if __name__ == '__main__':
         if policy[0].tool_name == '[done]':
             break
 
-        tool = ToolFactory(policy[0].tool_name, conf = load_config(f"{policy[0].tool_name}.yaml"))
+        tool = ToolFactory(policy[0].tool_name, conf=load_config(f"{policy[0].tool_name}.yaml"))
 
         observation, reward, terminated, _, info = tool.step(policy)
 
         print(observation)
 
     print(policy[0].policy_info)
-
-
