+++ aworld/core/common.py
@@ -16,36 +16,6 @@ class ActionResult(BaseModel):
     keep: bool = False
 
 
-class Tools(Enum):
-    """Tool list supported in the framework."""
-    BROWSER = "browser"
-    ANDROID = "android"
-    GYM = "openai_gym"
-    SEARCH_API = "search_api"
-    SHELL = "shell"
-    PYTHON_EXECUTE = "python_execute"
-    CODE_EXECUTE = "code_execute"
-    FILE = "file"
-    IMAGE_ANALYSIS = "image_analysis"
-    DOCUMENT_ANALYSIS = "document_analysis"
-
-
-class Agents(Enum):
-    """Tool list supported in the framework."""
-    BROWSER = "browser_agent"
-    ANDROID = "android_agent"
-    SEARCH = "search_agent"
-    CODE_EXECUTE = "code_execute_agent"
-    FILE = "file_agent"
-    IMAGE_ANALYSIS = "image_analysis_agent"
-    SHELL = "shell_agent"
-    DOCUMENT = "document_agent"
-    GYM = "gym_agent"
-    PLAN = "plan_agent"
-    EXECUTE = "execute_agent"
-    SUMMARY = "summary_agent"
-
-
 class Observation(BaseModel):
     """Observation information obtained from the tools.
