+++ aworld/virtual_environments/android/action/adb_actions.py
@@ -3,9 +3,10 @@
 
 import json
 
-from aworld.core.envs.tool_action import AndroidAction
+from aworld.config.common import Tools
+from aworld.config.tool_action import AndroidAction
 from aworld.core.envs.action_factory import ActionFactory
-from aworld.core.common import ActionModel, ActionResult, Tools
+from aworld.core.common import ActionModel, ActionResult
 from aworld.virtual_environments.android.action.adb_controller import ADBController
 from aworld.virtual_environments.android.config.android_action_space import AndroidActionParamEnum
 from aworld.virtual_environments.action import ExecutableAction
