+++ aworld/virtual_environments/android/android.py
@@ -4,8 +4,9 @@
 import traceback
 from typing import Any, Tuple, List, Dict
 
-from aworld.core.envs.tool_action import AndroidAction
-from aworld.core.common import ActionModel, Observation, ActionResult, Tools
+from aworld.config.common import Tools
+from aworld.config.tool_action import AndroidAction
+from aworld.core.common import ActionModel, Observation, ActionResult
 from aworld.logs.util import logger
 from aworld.virtual_environments.android.action.adb_controller import ADBController
 from aworld.virtual_environments.android.action.executor import AndroidToolActionExecutor
