+++ aworld/virtual_environments/apis/actions.py
@@ -7,9 +7,10 @@ import requests
 
 from typing import Tuple, Any, List, Dict
 
-from aworld.core.envs.tool_action import SearchAction
+from aworld.config.common import Tools
+from aworld.config.tool_action import SearchAction
 from aworld.core.envs.action_factory import ActionFactory
-from aworld.core.common import ActionModel, ActionResult, Tools
+from aworld.core.common import ActionModel, ActionResult
 from aworld.logs.util import logger
 from aworld.virtual_environments.action import ExecutableAction
