+++ aworld/virtual_environments/apis/search_api.py
@@ -4,10 +4,11 @@
 import json
 from typing import List, Tuple, Dict, Any
 
+from aworld.config.common import Tools
 from aworld.config.conf import ToolConfig
 from aworld.core.envs.tool import Tool, ToolFactory
-from aworld.core.envs.tool_action import SearchAction
-from aworld.core.common import Observation, ActionModel, Tools, ActionResult
+from aworld.config.tool_action import SearchAction
+from aworld.core.common import Observation, ActionModel, ActionResult
 from aworld.logs.util import logger
 from aworld.virtual_environments.utils import build_observation
