+++ aworld/virtual_environments/browsers/action/actions.py
@@ -9,9 +9,10 @@ from typing import Tuple, Any
 
 from langchain_core.prompts import PromptTemplate
 
-from aworld.core.envs.tool_action import BrowserAction
+from aworld.config.common import Tools
+from aworld.config.tool_action import BrowserAction
 from aworld.core.envs.action_factory import ActionFactory
-from aworld.core.common import ActionModel, ActionResult, Observation, Tools
+from aworld.core.common import ActionModel, ActionResult, Observation
 from aworld.virtual_environments.browsers.util.dom import DOMElementNode
 from aworld.logs.util import logger
 from aworld.virtual_environments.browsers.action.utils import DomUtil
