+++ aworld/virtual_environments/document/document.py
@@ -12,8 +12,9 @@ from urllib.parse import urlparse
 from pydantic import BaseModel
 
 from aworld.config import ToolConfig
-from aworld.core.envs.tool_action import DocumentExecuteAction
-from aworld.core.common import Tools, Observation, ActionModel, ActionResult
+from aworld.config.common import Tools
+from aworld.config.tool_action import DocumentExecuteAction
+from aworld.core.common import Observation, ActionModel, ActionResult
 from aworld.core.envs.tool import ToolFactory, Tool
 from aworld.logs.util import logger
 from aworld.virtual_environments.document.utils import encode_image_from_file, encode_image_from_url
