+++ aworld/virtual_environments/terminals/actions.py
@@ -1,9 +1,8 @@
 # coding: utf-8
 # Copyright (c) 2025 inclusionAI.
-
-from aworld.core.envs.tool_action import ShellAction
+from aworld.config.common import Tools
+from aworld.config.tool_action import ShellAction
 from aworld.core.envs.action_factory import ActionFactory
-from aworld.core.common import Tools
 from aworld.virtual_environments.action import ExecutableAction
