+++ aworld/virtual_environments/travel/writeTool.py
@@ -1,9 +1,9 @@
 # coding: utf-8
-import traceback
+import json
 from typing import Any, Tuple, List, Dict
 
-from aworld.core.envs.tool_action import WriteAction
-from aworld.core.common import ActionModel, Observation, ActionResult, Tools
+from aworld.config.tool_action import WriteAction
+from aworld.core.common import ActionModel, Observation, ActionResult
 from aworld.logs.util import logger
 from aworld.core.envs.tool import ToolFactory, Tool
 from aworld.config.conf import ToolConfig
